L-Variant/consumers/webhook/requirements.txt
confluent-kafka~=1.7.0
requests
kafka-python~=1.4.7
faust